First working content script
The content script replaces instances of "covid" with
"COVID" everywhere on washingtonpost.com  Still to come
is a UI that allows users to prefer that "covid" be
replaced with "COVID".

Known issue:  This will replace "covid" with "COVID" in
scripts as well as in text rendered to the user.  This
appears to be the state-of-the-art in text replacement
Chrome extensions.  I intend to continue searching for
a heuristic that reliably excludes script from
replacement.

Finally, there must be some way to avoid repeating myself
in the manifest and still cover both http and https
protocols for content script injection.